+++ common/src/main/java/generations/gg/generations/core/generationscore/common/world/item/tools/GenerationsSwordItem.java
@@ -18,8 +18,8 @@ import java.util.Set;
 public class GenerationsSwordItem extends SwordItem implements ToolEffectHolder<GenerationsSwordItem> {
     public final Set<ToolEffect> toolEffects = new HashSet<>();
 
-    public GenerationsSwordItem(Tier tier, int attackDamage, float attackSpeed, Properties properties) {
-        super(tier, attackDamage, attackSpeed, properties);
+    public GenerationsSwordItem(Tier tier, float attackDamage, float attackSpeed, Properties properties) {
+        super(tier, (int) attackDamage, attackSpeed, properties);
     }
 
     public GenerationsSwordItem addToolEffects(ToolEffect... toolEffect) {
